requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiosqlite>=0.19.0
